feat: add SyncPGMQueueOperationsMixin for synchronous queue operations
- Implemented a mixin class that provides shared synchronous operations for PGMQ.
- Included methods for queue management, message sending, topic-based routing, message reading, and metrics.
- Added support for transactional methods and JSON encoding.
- Enhanced logging and warnings for backward compatibility.
